shared default rt across tests
