+++ model/src/test/test_latent_admissions.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
 # numpydoc ignore=GL08
 
+from test.utils import get_default_rt
+
 import jax.numpy as jnp
 import numpy.testing as testing
 import numpyro
 import numpyro.distributions as dist
-from pyrenew import transformation as t
 from pyrenew.deterministic import DeterministicPMF
 from pyrenew.latent import HospitalAdmissions, Infections
-from pyrenew.metaclass import DistributionalRV, TransformedRandomVariable
-from pyrenew.process import SimpleRandomWalkProcess
+from pyrenew.metaclass import DistributionalRV
 
 
 def test_admissions_sample():
@@ -20,19 +20,7 @@ def test_admissions_sample():
 
     # Generating Rt and Infections to compute the hospital admissions
 
-    rt = TransformedRandomVariable(
-        name="Rt_rv",
-        base_rv=SimpleRandomWalkProcess(
-            name="log_rt",
-            step_rv=DistributionalRV(
-                name="rw_step_rv", dist=dist.Normal(0, 0.025)
-            ),
-            init_rv=DistributionalRV(
-                name="init_log_rt", dist=dist.Normal(0, 0.2)
-            ),
-        ),
-        transforms=t.ExpTransform(),
-    )
+    rt = get_default_rt()
 
     with numpyro.handlers.seed(rng_seed=223):
         sim_rt = rt(n_steps=30)[0].value
