+++ model/src/test/test_predictive.py
@@ -4,9 +4,10 @@
 Ensures that posterior predictive samples are not generated when no posterior samples are available.
 """
 
+from test.utils import get_default_rt
+
 import jax.numpy as jnp
 import numpyro.distributions as dist
-import pyrenew.transformation as t
 import pytest
 from pyrenew.deterministic import DeterministicPMF
 from pyrenew.latent import (
@@ -14,10 +15,9 @@ from pyrenew.latent import (
     Infections,
     InitializeInfectionsZeroPad,
 )
-from pyrenew.metaclass import DistributionalRV, TransformedRandomVariable
+from pyrenew.metaclass import DistributionalRV
 from pyrenew.model import RtInfectionsRenewalModel
 from pyrenew.observation import PoissonObservation
-from pyrenew.process import SimpleRandomWalkProcess
 
 pmf_array = jnp.array([0.25, 0.1, 0.2, 0.45])
 gen_int = DeterministicPMF(name="gen_int", value=pmf_array)
@@ -29,17 +29,7 @@ I0 = InfectionInitializationProcess(
 )
 latent_infections = Infections()
 observed_infections = PoissonObservation("poisson_rv")
-rt = TransformedRandomVariable(
-    "Rt_rv",
-    base_rv=SimpleRandomWalkProcess(
-        name="log_rt",
-        step_rv=DistributionalRV(
-            name="rw_step_rv", dist=dist.Normal(0, 0.025)
-        ),
-        init_rv=DistributionalRV(name="init_log_rt", dist=dist.Normal(0, 0.2)),
-    ),
-    transforms=t.ExpTransform(),
-)
+rt = get_default_rt()
 
 model = RtInfectionsRenewalModel(
     I0_rv=I0,
